+++ bukkit/src/main/java/de/maxhenkel/voicechat/command/VoiceChatCommands.java
@@ -29,9 +29,11 @@ public class VoiceChatCommands implements CommandExecutor {
         if (checkNoVoicechat(commandSender)) {
             return true;
         }
-        if (!(commandSender instanceof Player sender)) {
+        if (!(commandSender instanceof Player)) {
             return true;
         }
+        Player sender = (Player) commandSender;
+
         if (args.length >= 1) {
             if (args[0].equalsIgnoreCase("help")) {
                 return helpCommand(sender, command, label, args);
@@ -237,11 +239,11 @@ public class VoiceChatCommands implements CommandExecutor {
     }
 
     private static boolean checkNoVoicechat(CommandSender commandSender) {
-        if (commandSender instanceof Player player) {
-            if (Voicechat.SERVER.isCompatible(player)) {
+        if (commandSender instanceof Player) {
+            if (Voicechat.SERVER.isCompatible((Player) commandSender)) {
                 return false;
             }
-            commandSender.sendMessage(Voicechat.TRANSLATIONS.voicechatNeededForCommandMessage.get().formatted("Simple Voice Chat"));
+            commandSender.sendMessage(String.format(Voicechat.TRANSLATIONS.voicechatNeededForCommandMessage.get(), "Simple Voice Chat"));
         } else {
             commandSender.sendMessage(Voicechat.TRANSLATIONS.playerCommandMessage.get());
         }